+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Welcome to the Github! If you are looking for the most recent stable version, th
 ##### Block:
 - World Blender Portal block rendering is now optimized thanks to comp500! Special thanks to them!
 
+- Added screen overlay when inside the portal block.
+
 ##### Misc:
 - Fixed several mixins not being clientside when they should be.